refactor(plugin): generate integration types
