+++ packages/plugin/src/v2/effect/index.ts
@@ -1,8 +1,3 @@
 export type { PluginContext } from "./context.js"
 export { define } from "./plugin.js"
 export type { Plugin } from "./plugin.js"
-export type {
-  IntegrationCredential,
-  IntegrationOAuthCredential,
-  IntegrationOAuthMethodRegistration,
-} from "./integration.js"
