+++ packages/plugin/src/v2/effect/skill.ts
@@ -1,14 +1,9 @@
-import type { SkillV2Info } from "@opencode-ai/sdk/v2/types"
+import type { SkillV2Source } from "@opencode-ai/sdk/v2/types"
 import type { Hooks } from "./registration.js"
 
-export type SkillSource =
-  | { readonly type: "directory"; readonly path: string }
-  | { readonly type: "url"; readonly url: string }
-  | { readonly type: "embedded"; readonly skill: SkillV2Info }
-
 export interface SkillDraft {
-  source(source: SkillSource): void
-  list(): readonly SkillSource[]
+  source(source: SkillV2Source): void
+  list(): readonly SkillV2Source[]
 }
 
 export type SkillHooks = Hooks<{
